requirements.txt
fastapi
uvicorn[standard]
psycopg2
python-dotenv

tensorflow
keras
Pillow